Move Failable into its own module
